+++ jason/src/lib.rs
@@ -29,6 +29,7 @@ pub use console_error_panic_hook::set_once as set_panic_hook;
 pub use self::{
     api::{ConnectionHandle, Jason, RoomHandle},
     media::{
+        track::{local::JsTrack, remote::Track},
         AudioTrackConstraints, DeviceVideoTrackConstraints,
         DisplayVideoTrackConstraints, FacingMode, JsMediaSourceKind, MediaKind,
         MediaStreamSettings,
